tail: fix pipe-f test by detecting broken stdout pipe
