+++ src/uu/tee/Cargo.toml
@@ -19,7 +19,6 @@ path = "src/tee.rs"
 
 [dependencies]
 clap = { workspace = true }
-nix = { workspace = true, features = ["poll", "fs"] }
 uucore = { workspace = true, features = ["libc", "parser", "signals"] }
 fluent = { workspace = true }
