+++ src/uu/tee/src/tee.rs
@@ -3,8 +3,6 @@
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
 
-// cSpell:ignore POLLERR POLLRDBAND pfds revents
-
 use clap::{Arg, ArgAction, Command, builder::PossibleValue};
 use std::ffi::OsString;
 use std::fs::OpenOptions;
@@ -18,6 +16,8 @@ use uucore::{format_usage, show_error};
 
 // spell-checker:ignore nopipe
 
+#[cfg(target_os = "linux")]
+use uucore::signals::ensure_stdout_not_broken;
 #[cfg(unix)]
 use uucore::signals::{enable_pipe_errors, ignore_interrupts};
 
@@ -421,46 +421,4 @@ impl Read for NamedReader {
             okay => okay,
         }
     }
-}
-
-/// Check that if stdout is a pipe, it is not broken.
-#[cfg(target_os = "linux")]
-pub fn ensure_stdout_not_broken() -> Result<bool> {
-    use nix::{
-        poll::{PollFd, PollFlags, PollTimeout},
-        sys::stat::{SFlag, fstat},
-    };
-    use std::os::fd::AsFd;
-
-    let out = stdout();
-
-    // First, check that stdout is a fifo and return true if it's not the case
-    let stat = fstat(out.as_fd())?;
-    if !SFlag::from_bits_truncate(stat.st_mode).contains(SFlag::S_IFIFO) {
-        return Ok(true);
-    }
-
-    // POLLRDBAND is the flag used by GNU tee.
-    let mut pfds = [PollFd::new(out.as_fd(), PollFlags::POLLRDBAND)];
-
-    // Then, ensure that the pipe is not broken.
-    // Use ZERO timeout to return immediately - we just want to check the current state.
-    let res = nix::poll::poll(&mut pfds, PollTimeout::ZERO)?;
-
-    if res > 0 {
-        // poll returned with events ready - check if POLLERR is set (pipe broken)
-        let error = pfds.iter().any(|pfd| {
-            if let Some(revents) = pfd.revents() {
-                revents.contains(PollFlags::POLLERR)
-            } else {
-                true
-            }
-        });
-        return Ok(!error);
-    }
-
-    // res == 0 means no events ready (timeout reached immediately with ZERO timeout).
-    // This means the pipe is healthy (not broken).
-    // res < 0 would be an error, but nix returns Err in that case.
-    Ok(true)
 }
